requirements.txt
beautifulsoup4==4.6.3
alphabet_detector
gTTS
jikanpy
psutil
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2
feedparser
pythonping
pynewtonmath
spongemock
zalgo-text
geopy
nltk
aiohttp>=2.2.5
Pillow>=4.2.0
PyLyrics
hurry.filesize
googletrans
tldextract
PyYAML
pyowm
wikipedia
coffeehouse
bleach
markdown2
speedtest-cli
